Add support for .vue single-file components
