--- assets/js/vue-components/vue-config.js
@@ -1,7 +1,0 @@
-export const VueConfig = {
-    delimiters: ['<%', '%>']
-};
-
-export function withDefaults(config) {
-    return Object.assign({}, VueConfig, config);
-}
