+++ package.json
@@ -125,7 +125,7 @@
     "swiper": "^4.1.6",
     "uuid": "^3.1.0",
     "vary": "^1.1.1",
-    "vue": "^2.4.4",
+    "vue": "2.5.16",
     "yargs": "^11.0.0"
   },
   "devDependencies": {
@@ -133,7 +133,7 @@
     "autotrack": "^2.4.1",
     "babel-core": "^6.26.0",
     "babel-eslint": "^8.0.1",
-    "babel-loader": "^7.1.2",
+    "babel-loader": "^7.1.4",
     "babel-plugin-lodash": "^3.2.11",
     "babel-plugin-syntax-dynamic-import": "^6.18.0",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
@@ -141,6 +141,7 @@
     "babel-preset-env": "^1.6.0",
     "chai": "^4.1.2",
     "concurrently": "^3.5.0",
+    "css-loader": "^0.28.11",
     "cssnano": "^3.10.0",
     "cypress": "^3.0.1",
     "del": "^3.0.0",
@@ -169,6 +170,8 @@
     "pretty": "^2.0.0",
     "prompt": "^1.0.0",
     "run-sequence": "^2.2.0",
+    "vue-loader": "^15.2.4",
+    "vue-template-compiler": "2.5.16",
     "wait-on": "^2.1.0",
     "webpack": "^4.0.1",
     "webpack-cli": "^2.0.9"
