+++ views/pages/funding/order-free-materials.njk
@@ -126,7 +126,7 @@
         </div>
     {% else %}
         {{ super() }}
-        <div id="js-vue" class="no-vue" data-orders="{{ orders | dump }}">
+        <div id="js-materials" class="no-vue" data-orders="{{ orders | dump }}">
             <div class="content-box inner--wide-only accent--{{ pageAccent }} a--border-top">
                 <h2>{{ copy.locationPrompt }}</h2>
                 <ul class="unstyled grid grid--padded grid--equal grid--2-up grid--wide-only">
