+++ webpack.config.js
@@ -1,11 +1,11 @@
 /* eslint-env node */
 'use strict';
-
-const webpack = require('webpack');
-
 const path = require('path');
 const pkg = require('./package.json');
 const { getBuildSummary } = require('./build-helpers');
+
+const webpack = require('webpack');
+const VueLoaderPlugin = require('vue-loader/lib/plugin');
 
 const buildSummary = getBuildSummary();
 
@@ -19,9 +19,11 @@ const commonConfig = {
             {
                 test: /\.js$/,
                 exclude: /node_modules/,
-                use: {
-                    loader: 'babel-loader'
-                }
+                loader: 'babel-loader'
+            },
+            {
+                test: /\.vue$/,
+                loader: 'vue-loader'
             }
         ]
     }
@@ -55,6 +57,7 @@ module.exports = [
             }
         },
         plugins: [
+            new VueLoaderPlugin(),
             new webpack.BannerPlugin({
                 banner: `${pkg.description} - ${buildSummary.commitHash}`
             })
